matslise3d split sector in different compile unit
