+++ test/matslise3d/coulomb.cpp
@@ -7,15 +7,20 @@ using namespace std;
 
 TEST_CASE("3D: coulomb", "[matslise3d][coulomb]") {
     // https://arxiv.org/pdf/0904.0939.pdf
-/*
+
+    Matslise3D<>::Config config;
+    config.tolerance = 1e-7;
+    config.zStepsPerSector = 5;
+    config.zSectorBuilder = sector_builder::uniform<Matslise3D<>>(16);
+    config.ySectorBuilder = sector_builder::automatic<Matslise2D<>>(1e-6);
+
     const double a = 0.05;
     Matslise3D<double> matslise([a](double x, double y, double z) -> double {
         double r = hypot(hypot(x, y), z);
         if (r < a)
             return -2 / a;
         return -2 / r;
-    }, {{{-12.3, 12.3}, -12.3, 12.3}, -12.3, 12.3}, sector_builder::uniform<Matslise3D<double>>(22), 1e-7);
+    }, {{{-12.3, 12.3}, -12.3, 12.3}, -12.3, 12.3}, config);
 
     cout << matslise.eigenvalue(-1) << endl;
-    */
 }
